requirements.txt
Pillow==10.3.0
tkintermapview==1.27
pandas==2.2.2
bcrypt==4.1.3
python-dotenv==1.0.1
